+++ vec-common/src/main/java/com/foursoft/harness/vec/common/HasIdentification.java
@@ -23,23 +23,11 @@
  * THE SOFTWARE.
  * =========================LICENSE_END==================================
  */
-package com.foursoft.vecmodel.common.util;
+package com.foursoft.harness.vec.common;
 
-import java.util.List;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
+@FunctionalInterface
+public interface HasIdentification {
 
-public class DelegationUtils {
-
-    public static <T extends X, X> List<T> getFromListWithType(final List<X> source, final Class<T> type) {
-        return getFromListWithTypeAsStream(source, type)
-                .collect(Collectors.toList());
-    }
-
-    public static <T extends X, X> Stream<T> getFromListWithTypeAsStream(final List<X> source, final Class<T> type) {
-        return source.stream()
-                .filter(type::isInstance)
-                .map(type::cast);
-    }
+    String getIdentification();
 
 }
