+++ vec120/src/main/java/com/foursoft/harness/vec/v12x/navigations/PartOccurrenceOrUsageNavs.java
@@ -27,9 +27,9 @@ package com.foursoft.harness.vec.v12x.navigations;
 
 import com.foursoft.harness.vec.v12x.*;
 import com.foursoft.harness.vec.v12x.visitor.ReferencedNodeLocationVisitor;
-import com.foursoft.vecmodel.common.annotations.RequiresBackReferences;
-import com.foursoft.vecmodel.common.util.StreamUtils;
-import com.foursoft.vecmodel.common.util.StringUtils;
+import com.foursoft.harness.vec.common.annotations.RequiresBackReferences;
+import com.foursoft.harness.vec.common.util.StreamUtils;
+import com.foursoft.harness.vec.common.util.StringUtils;
 
 import java.util.Collection;
 import java.util.Optional;
